Move the coarse-pointer debounce test to its own isolated file
The coarse-pointer case stubs window and resets the module registry to
re-import the effects module with a touch pointer class. vitest.shared.ts
treats vi.stubGlobal/vi.resetModules as isolation markers, so that one
test moved the whole 65-test realtime-cache-effects suite out of the
shared node worker into its own isolated worker, paying the module-graph
re-import for every run. Keep the case in a small dedicated file so only
it runs isolated and the main suite returns to the shared worker.
